# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CITMS 3.6 API Documentation
  description: |
    Centralized IT Management System (CITMS) v3.6 API.
    This API handles IT Asset Management, Inventory Ingestion, License Compliance, ITSM, Remote Control, Procurement, and Workflow Automation.
    
    All error responses follow the **RFC 7807 (Problem Details for HTTP APIs)** standard.
  version: "3.6.0"
  contact:
    name: CITMS Support
    email: support@citms.com
servers:
  - url: https://api.citms.com/api/v1
    description: Production Server
  - url: http://localhost:8000/api/v1
    description: Local Development Server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Device:
      type: object
      properties:
        id:
          type: string
          format: uuid
        hostname:
          type: string
        serial_number:
          type: string
        device_type:
          type: string
          enum: [WORKSTATION, SERVER, LAPTOP, NETWORK, MOBILE]
        status:
          type: string
          enum: [ONLINE, OFFLINE, MAINTENANCE, DISPOSED]
        ip_address:
          type: string
        mac_address:
          type: string
        last_seen:
          type: string
          format: date-time

    # ITSM Schemas
...
paths:
...
  /devices:
    get:
      tags: [Asset Management]
      summary: List all devices
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
  /devices/{id}:
    get:
      tags: [Asset Management]
      summary: Get device details
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Device details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'

  # --- INVENTORY ---
...